--- www/api/getCurrentDayTimeSpent.php
@@ -1,4 +1,0 @@
-<?php
-// Deprecated!
-print_r('{"DaySum":"0"}');
-?>
