+++ www/api/habits.php
@@ -26,16 +26,17 @@ if (isset($_GET['ID'])) {
   
   print '{"status": "done", "sql": "' . $sql . '"}';
 } else {
-  $sql = "SELECT habits.ID, habits.name, habits.type, GROUP_CONCAT(DATE(done)) AS dates FROM `habits` LEFT JOIN habits_tracker ON habits.ID = habits_tracker.habitID WHERE habits.active = 1 GROUP BY habits.ID;";
+  $sql = "SELECT habits.ID, habits.name, DATE(habits.created) as created, habits.type, GROUP_CONCAT(DATE(done)) AS dates FROM `habits` LEFT JOIN habits_tracker ON habits.ID = habits_tracker.habitID WHERE habits.active = 1 GROUP BY habits.ID;";
   $result = mysqli_query($db, $sql);
 
   $rows = array();
   while ($row = mysqli_fetch_assoc($result)) {
     $id = $row['ID'];
     $name = $row['name'];
+    $created = $row['created'];
     $type = $row['type'];
     $dates = explode(",", $row['dates']);
-    $rows[] = ["id" => $id, "name" => $name, "type" => $type, "dates" => $dates];
+    $rows[] = ["id" => $id, "name" => $name, "created" => $created, "type" => $type, "dates" => $dates];
   }
 
   header('Content-Type: application/json');
