--- www/js/config.js
@@ -1,3 +1,0 @@
-const config = {
-  debug: false,
-}
