+++ www/js/habits.js
@@ -20,7 +20,7 @@ function getPrevMonth(date) {
 }
 
 function createTable(table, month, habits) {
-    while (table.hasChildNodes()) {
+    while (table?.hasChildNodes()) {
         table.removeChild(table.lastChild);
     }
 
@@ -54,13 +54,15 @@ function createTable(table, month, habits) {
         tr.appendChild(title)
 
         for (let i = 1; i <= days_in_month; i += 1) {
-            const date = habits_month.getFullYear() + "-" + (habits_month.getMonth() + 1) + "-" + i.toString().padStart(2, "0")
+            const date = habits_month.getFullYear() + "-" + (habits_month.getMonth() + 1).toString().padStart(2, "0") + "-" + i.toString().padStart(2, "0")
             const td = document.createElement("td")
-            const input = document.createElement("input")
-            input.type = "checkbox"
-            input.addEventListener("click", event => toggleHabit(event, habit.id, date))
-            input.checked = habit.dates.includes(date)
-            td.appendChild(input)
+            if((new Date(habit.created)) <= (new Date(date)) || habit.dates.includes(date)) {
+                const input = document.createElement("input")
+                input.type = "checkbox"
+                input.addEventListener("click", event => toggleHabit(event, habit.id, date))
+                input.checked = habit.dates.includes(date)
+                td.appendChild(input)
+            }
             tr.appendChild(td)
         }
 
@@ -81,21 +83,23 @@ const habits_table_monthly = document.getElementById("habits_table_monthly")
 const habits_next_month = document.getElementById("habits_next_month")
 const habits_prev_month = document.getElementById("habits_prev_month")
 
-habits_next_month.addEventListener("click", () => { habits_month = getNextMonth(habits_month); createHabitTables() })
-habits_prev_month.addEventListener("click", () => { habits_month = getPrevMonth(habits_month); createHabitTables() })
+habits_next_month?.addEventListener("click", () => { habits_month = getNextMonth(habits_month); createHabitTables() })
+habits_prev_month?.addEventListener("click", () => { habits_month = getPrevMonth(habits_month); createHabitTables() })
 
 async function createHabitTables() {
     let habits = await fetchHabits()
 
     const daily = habits.filter(x => x.type === "daily")
-    const weekly = habits.filter(x => x.type === "weekly")
-    const monthly = habits.filter(x => x.type === "monthly")
+    // const weekly = habits.filter(x => x.type === "weekly")
+    // const monthly = habits.filter(x => x.type === "monthly")
 
     createTable(habits_table_daily, habits_month, daily)
     // createTable(habits_table_weekly, habits_month, weekly)
     // createTable(habits_table_monthly, habits_month, monthly)
 
-    document.getElementById("habits_curr_date").innerText = (habits_month.getMonth() + 1) + "-" + habits_month.getFullYear()
+    document.getElementById("habits_curr_date").innerText = (habits_month.getMonth() + 1).toString().padStart(2, "0") + "-" + habits_month.getFullYear()
 }
 
-createHabitTables()
+if(habits_table_daily) {
+    createHabitTables()
+}
